Fix dependency violations
