+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/project/api/queries.kt
@@ -1,7 +1,5 @@
 package com.novatecgmbh.eventsourcing.axon.project.project.api
 
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
-
 class AllProjectsQuery
 
 data class ProjectQuery(val projectId: ProjectId)
