+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/project/api/value_objects.kt
@@ -1,0 +1,13 @@
+package com.novatecgmbh.eventsourcing.axon.project.project.api
+
+import com.fasterxml.jackson.annotation.JsonValue
+import java.io.Serializable
+import java.util.*
+import javax.persistence.Embeddable
+
+@Embeddable
+data class ProjectId(@get:JsonValue val identifier: String) : Serializable {
+  constructor() : this(UUID.randomUUID().toString())
+
+  override fun toString(): String = identifier
+}
