+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/project/query/ProjectProjectionRepository.kt
@@ -1,7 +1,7 @@
 package com.novatecgmbh.eventsourcing.axon.project.project.query
 
 import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectProjection
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
+import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectId
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
