+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/project/web/ProjectControllerV1.kt
@@ -1,11 +1,6 @@
 package com.novatecgmbh.eventsourcing.axon.project.project.web
 
-import com.novatecgmbh.eventsourcing.axon.project.project.api.AllProjectsQuery
-import com.novatecgmbh.eventsourcing.axon.project.project.api.CreateProjectCommand
-import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectQuery
-import com.novatecgmbh.eventsourcing.axon.project.project.api.UpdateProjectCommand
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
-import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectProjection
+import com.novatecgmbh.eventsourcing.axon.project.project.api.*
 import com.novatecgmbh.eventsourcing.axon.project.project.web.dto.CreateProjectDto
 import java.time.Duration
 import java.time.LocalDate
