+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/project/web/ProjectControllerV2.kt
@@ -1,8 +1,6 @@
 package com.novatecgmbh.eventsourcing.axon.project.project.web
 
 import com.novatecgmbh.eventsourcing.axon.project.project.api.*
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
-import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectProjection
 import com.novatecgmbh.eventsourcing.axon.project.project.web.dto.CreateProjectDto
 import com.novatecgmbh.eventsourcing.axon.project.project.web.dto.RenameProjectDto
 import com.novatecgmbh.eventsourcing.axon.project.project.web.dto.RescheduleProjectDto
