+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/task/api/queries.kt
@@ -1,7 +1,6 @@
 package com.novatecgmbh.eventsourcing.axon.project.task.api
 
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
-import com.novatecgmbh.eventsourcing.axon.project.task.command.TaskId
+import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectId
 
 data class TasksByProjectQuery(val projectId: ProjectId)
