+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/task/query/TaskProjectionRepository.kt
@@ -1,7 +1,8 @@
 package com.novatecgmbh.eventsourcing.axon.project.task.query
 
-import com.novatecgmbh.eventsourcing.axon.project.project.command.ProjectId
-import com.novatecgmbh.eventsourcing.axon.project.task.command.TaskId
+import com.novatecgmbh.eventsourcing.axon.project.project.api.ProjectId
+import com.novatecgmbh.eventsourcing.axon.project.task.api.TaskId
+import com.novatecgmbh.eventsourcing.axon.project.task.api.TaskProjection
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
