+++ src/main/kotlin/com/novatecgmbh/eventsourcing/axon/project/task/web/dto/RescheduleTaskDto.kt
@@ -1,7 +1,7 @@
 package com.novatecgmbh.eventsourcing.axon.project.task.web.dto
 
-import RescheduleTaskCommand
-import com.novatecgmbh.eventsourcing.axon.project.task.command.TaskId
+import com.novatecgmbh.eventsourcing.axon.project.task.api.RescheduleTaskCommand
+import com.novatecgmbh.eventsourcing.axon.project.task.api.TaskId
 import java.time.LocalDate
 
 data class RescheduleTaskDto(val startDate: LocalDate, val endDate: LocalDate) {
